Update dependencies and remove Chart.js dependency
This commit updates all our dependencies to their latest version and
completely remove Chart.js as a dependency.

Chart.js was (only) used to generate the "BufferSize" graph in the demo
page. However, it went through a new major version (Chart.js 3) which
wasn't compatible with the old syntax.

I tried to do the migration by following their "migration guide" yet
failed, then tried to re-read their documentation from start to
re-implement it from scratch but still failed for the same reason: an
incomprehensible error completely broke the page when I wanted to show
the graph.

I lost patience and just decided to remove the dependency - and the
BufferSize chart with it. How we were using it (animating it by
re-rendering a new one each time on top) was not optimal anyway.

Re-implementing the BufferSize chart with canvas might not be the
hardest thing to do and I think we should try it.
We do already create the other "BufferContent" chart ourselves through a
canvas, and they seem to amount to the same level of complexity.