Demo: re-implement buffer size chart
Based on #955.

This commit re-implement a buffer size chart from scratch in the demo now
that the last dependency update ditched out `chart.js`.

It is still a little rough for now, but I find it very usable. It was
also very fun to do and might be also fun to improve when we find the
time.

Thoughts?