dash: do not trigger refresh from a ListRepresentationIndex
`RepresentationIndex`es, which are basically interfaces to get
information about segment regardless of the current streaming protocol,
have a function called `shouldRefresh`

This function takes in argument the time range currently considered for
download by the player and returns `true` if the `RepresentationIndex`
estimates that a new Manifest has more information on the wanted segments
and therefore that it should be refreshed.

Used for both DASH and Smooth at first (years ago?), it finally was
considered only useful for Smooth contents, as DASH also provide other
mechanisms for recurrent MPD updates that we now prefer (most notably
the `minimumUpdatePeriod` attribute) over this.

However a logic was still present for `ListRepresentationIndex` - used
when a `<SegmentList> element is encountered in a MPD - despite the fact
that they're (almost?) always for static contents.

There `shouldRefresh` would return `true` if the wanted time range went
over the end of the last segment in the SegmentList. This seems fine, yet
it can return false positives in cases of not-perfectly formed MPDs and
rounding errors.

The result would be unnecessary new MPD requests.

I thus chose to always return `false` for that method, like for any
other `shouldRefresh` implementation beside the smooth one.

This bug was seen as part of the investigation for the #960 issue.