wasm/dash: handle Event namespaces
There was a subtle issue with the WebAssembly DASH parser, which could
lead to inexploitable `streamEvent` events (emit when an `<Event>` element
encountered in the MPD begin to be "active").

As per the RxPlayer API documentation, `<Event> elements are sent as an
`Element` JavaScript type through the `streamEvent` event, meaning that we
have to use the the DOM-related JavaScript APIs (to be able to create
that `Element` type).

However, the main API we used, DOMParser, performed validity checks on
the given XML.
We initially though that just feeding it the `<Event>` element was OK,
it was valid XML after all, but we forgot about XML namespaces!

Namespaces are special attributes and Element names. They are recognized
by the inclusion of the ":" character after them. They have to be
declared thanks to an `xmns:<NAMESPACE_NAME>` attribute in one of its
parent element.

As only the `<Event>` tag subset is given to it, not the full MPD (for
performance reasons as that API was one of the main reasons for us
switching to WASM). This means that namespaces declared elsewhere in the
MPD was not available to the DOMParser.

When the DOMParser API encounter an unknown namespace, it sadly fails.
This is the source of the issue I'm trying to fix.

---

After iterating through multiple other uglier but simpler solutions
(whose goal were mostly to work-around JavaScript validity checks), I
did not find any way of handling that properly (and surely).

As a last stretch, I tried to implement it more "cleanly", by:

  - registering all namespaces declaed in all `<Event>` element's
    parents (which are in order: the `<MPD>` element, the `<Period>`
    element and the `<EventStream>` element).

  - temporarily creating a parent element named `<toremove>` to that
    `<Event>`, just before feeding it to the DOMParser. This element
    will contain all declared namespaces in all three of its parents.

  - removing the `<toremove>` element just after the DOMParser API has
    been called.

This commit is the result of that experimentation. It seems to work
well with all tested contents.

Thoughts?