[WIP] Avoid requesting in a loop segments the browser crops for some reasons
This tries to fix #987.

In that issue, Safari advertised that the first pushed audio segment
started at 0.450+ where the RxPlayer previously thought it started at
exactly 0 second (based on MPD information and container metadata, such
as the `tfdt` box of ISOBMFF segments).

The behavior of the player when encountering such a big difference, is
to re-download the corresponding segment, because it thinks that the
pushed segment has been partly garbage-collected by the browser.

In Safari re-pushing the same segments lead to the same result and as
such we end up requesting the same segment in a loop while rebuffering
endlessly.

In that situation, it appears obvious that we should just ignore the
issue and just seek over that gap and continue. However having a
garbage-collected segment (even just after being pushed) could be a
legitimate case which might resolve itself by re-downloading the
segment.

So it's not too clear what the logic should be.

---

In this commit, I implement the following logic:

  1. I added an "history" abstraction on top of the SegmentInventory,
     which will record important information observed on buffered
     segments over time.

     For the moment, it only stores the initially observed buffered start
     time and end time of every pushed segments relatively to the
     expected start and end, even if those segments have since been
     replaced / removed.

     Because just storing an infinite history of pushed segments could
     lead to a consequent memory usage after long sessions, I put a maximal
     size (for the moment, 200 elements, which seems a lot more than
     enough here) and a maximum lifetime (every recorded events which
     are older than a minute are automatically removed).

  2. In the RepresentationStream, when it notices that an already
     buffered segment has been partially garbage collected at the start
     and/or end, the player checks thanks to the history if this same
     segment has already been pushed before and if it was, if we were
     in this same situation the last time it happened.

     If that's the case, the player does not re-ask for the segment to
     be loaded.

   3. Because the newer discontinuity-handling logic is well-designed
      (:p), the player is then able to automatically skip over the
      corresponding gap it knows will not be filled.

This is basically an heuristic which could be simplified as:
Don't retry to load segment which appears to be twice in a row incomplete
just after pushing it.

---

The history solution might appear weird and over-engineered when
compared to other more ad-hoc and minimal ones (for example, we could
have added some kind of tag to a re-downloaded segment instead and use
that tag to know that we can stop retrying it), but I still prefer it
for several reasons:

  - it seems easier to comprehend and maintain: the notion of an
    history is simple, prevalent and logical enough for it to find its
    place in both the SegmentBuffers module and the RepresentationStream
    one.

  - it allows to answer several other linked issues, such as segment
    partially GCed or even removed when on the border of other Period
    (which we don't much have anymore usually, due to other improvements,
    but which can still technically happen).

  - it also can help a lot during debugging sessions of those kind of
    issues.
    Knowing what happened in the buffers not only now but also in the
    past is a very nice tool to have.

    Today, when we debug those issues, we usually dive deep in debug
    logs and put breakpoints in specific loops iteration.
    Being able to just easily rely on an history of what happened
    chronologically is much better.

---

The code is still rough but it should work well.