stream: add the concept of a "locked" stream
-- Overview --

This commit fixes a rare infinite rebuffering issue occuring when
playing close - but before - the next Period's start, when one of
its associated audio and/or video PeriodStream or AdaptationStream
requires a reloading of the MediaSource before having the chance to
load any segment.

For example this can happen when disabling the video track of the next
Period, before it had the chance to load any segment. In that case, when
playing close to that Period, the clock will anounce that we should
rebuffer because there is not enough data ahead in the video buffer
to guarantee a smooth playback.

Yet the future video stream does not want to download any video segments
- because video tracks have been disabled - yet the `StreamOrchestrator`
refuses to ask for the removal of the video SourceBuffer - because we're
still playing the previous Period which needs a video SourceBuffer.

So this is kind of a fancy deadlock:
  - The init is putting the player in REBUFFERING mode while playing the
    previous Period, waiting for the next Period to load segments before
    playing it.
  - The stream awaits the time when the next Period is actively playing
    to be able to remove the video SourceBuffer so we can quit the
    rebuffering situation

Here I'm specifically talking about video SourceBuffer removal, but this
situation can ultimately happen anytime some stream refuse to load a
segment until it is the current one (which for the moment only includes
MediaSource reloading requests).

-- Solution implemented here

The current solution is sadly not perfect, yet it is simple enough and
should work in all possible cases.

First it introduces the concept of a "locked stream", which is sent by
a Stream when it refuses to download a segment until the Period is the
current one.
When a PeriodStream or AdaptationStream is left in such situation, it
looks if it seems to be the currently played Period:
  - if it is, it asked to reload the MediaSource instead
  - if it isn't, it instead sends a new "locked-stream" event

The `StallAvoider`, which is the part in the `Init` avoiding any
stalling possibilities, will then listen to those "locked-stream"
events.
If it is rebuffering waiting on data at least very close to that next
Period, it will automatically perform a seek to that Period's beginning.

-- Limitations with this solution

There are still multiple issues with this solution.

First, this moves the responsibility of checking which Period is the
current one from the `StreamOrchestrator` to the concerned
`PeriodStream` or `AdaptationStream`, which is arguably less apropriate.

Moving back the logic the `StreamOrchestrator` should be possible but I
did not look too much into it for now.

Also, due to how the RxPlayer is currently architectured, there isn't
any event to indicate that a stream has "unlocked".

This is very problematic because it means that the `StallAvoider` isn't
just able to store the last state for each Period and buffer type
combination (a simpler solution).
Any handling of locked stream needs thus to be done immediately after
the event is received. After, it won't know if it hasn't unlocked since
then.

The previous issue is also the reason why we have to seek at the
beginning of the next Period, instead of just temporarily disabling
RxPlayer's rebuffering logic to let regular playback reach the end of the
current buffer (which may already lead to the second Period) by itself.

This means that some frames will be unnecessarily skipped, which is also
a little sad :/.