Refacto the "clock" module into a more self-describing "PlaybackObserver"
The clock was initially named that way because it was originally an
Observable emitting playback conditions at regular intervals.

Hence this module was called the Clock (or `clock$`) due to its behavior,
and the playback conditions it emitted were called `ticks`, because
that's what clocks do.

That's somewhat cute but sadly we had to disfigure the poor thing in
recent times.

The last nail on its coffin was the `internal seek` feature, which
necessitated to add seeking capabilities to it - which kind of looked
out of place there. This was not the only bad thing we inflicted to that
poor clock though. We also had added a `getCurrentTime` function to each
of those ticks, so we would not just be able to obtain the position at
the time of the tick, but at any time we which.

So I thought that it was time to stop this masquerade and retire it.

This commit try to replace it by a now-more appropriately named
`PlaybackObserver`, which is now a class with methods such as:

  - getCurrentTime
  - setCurrentTime
  - listen (equivalent of the old clock's Observable)

There are still some rough edges though. First, the `tick` variable is
now more generally named `observation`, which is kind of a too-generic
(and kind of not nice-sounding) name.

Also the whole concept of `IReadOnlyPlaybackObserver` which are derived
PlaybackObserver which cannot influence directly playback information
(e.g. by seeking) can be complicated to understand at first.