dash: provide default startNumber for number-based SegmentTimeline
DASH indexes of segments which rely on a `<SegmentTemplate>` element can
either declare segment URLs based on the corresponding segment's
timestamp or on a number incrementing with each segment from a given
base, the `startNumber`.

Usually, time-based SegmentTemplates contains a `<SegmentTimeline>`
element, listing all segments and their corresponding timestamps whereas
number-based don't.

But this is not always the case, for example a number-based indexing
scheme can be linked to a `<SegmentTimeline>` in the MPD (we might even
in many case prefer it because a SegmentTimeline facilitates live-edge
detection on the player-side).

As per the DASH-IF IOP the `startNumber` is totally optional and should
be set to `1` by default.
This is something we considered when no SegmentTimeline was present yet
not when it was.

This is because the logic inside the RxPlayer is different in both
situations:
  - in the presence of a SegmentTimeline the RxPlayer's
    `SegmentTimelineRepresentationIndex` structure is used to retrieve
    segment URLs
  - in a SegmentTemplate without any SegmentTimeline, the RxPlayer uses
    a `SegmentTemplateRepresentationIndex`, which has a whole other logic

The implementation of a default value for `startNumber` was only
implemented for the latter. Consequently, any number-based index with a
SegmentTimeline but without a `startNumber` attribute would lead to an
error when obtaining the URL for the first needed segments.

This commit fixes that.

Note that the majority of the code here is to add integration tests
which tests that parsing a MPD with a SegmentTimeline and number-based
works well.