Add SharedReference object replacing most BehaviorSubjects
This commit is a proposal to use a new type of object, the
`ISharedReference`. It could be used when we want a value:

  - both settable and gettable by one part of the code, but only
    readable by much of the rest.

  - both gettable synchronously or subscribable to be notified of
    future updates.

Previously, we used a `BehaviorSubject` exactly for that - frequent -
usecase. However, I found that it has some drawbacks:

  - the main one, which recently bit me, is that a `BehaviorSubject` is
    always both readable (synchronously through `getValue` and
    asynchronously through subscription) and writable (through e.g.
    `next`, `error` and `complete`).

    This is problematic when we want to enforce and document through
    typing that a particular input should only be read into, but never
    mutated.

    For example, both the `Init` and various `Stream` modules received
    the `wantedBufferAhead$` value as a `BehaviorSubject`, despite the
    fact that they should NEVER modify it. This was to explicitely allow
    synchronous reads.

    The potential fix to that issue was previously to upcast the
    `BehaviorSubject` into an `Observable` instead.
    This works, but in doing that the notion of getting the last value
    the `BehaviorSubject` was set to is lost.

    An Observable derived from a `BehaviorSubject` will still
    synchronously emit its current value first when subscribed to, but
    this is not something enforced by either the type or the Observable's
    documentation. If we wanted to make it clear, we would have to
    document manually that we we need this behavior in the corresponding
    module's documentation.

    This is the reason why some values like `wantedBufferAhead$` still
    were typed as `BehaviorSubject` despite being read-only: a
    `BehaviorSubject` makes it clear that the initial value is returned.

    In the last release, it was even the main reason why we do not check
    the wanted playback rate to know when to unlock a "locked stream".
    Here we wanted to check Init's `speed$`'s value synchronously,
    without making that `Observable` a `BehaviorSubject` (as it is
    originally), because only the `API` should be able to update the
    wanted speed.

  - A `BehaviorSubject` is documented as: "A variant of Subject that
    requires an initial value and emits its current value whenever it is
    subscribed to."

    What we really want in most of our cases is just a shared value
    whose updates can be subscribed to, that's all.

    It is a slightly simpler concept and does not need all the
    agage behind `Subject`s, `BehaviorSubject`s or even `Observable`s
    (like errors, hot and cold and all the complex theory).

---

The idea in this commit is to implement a `ISharedReference` interface,
which has four methods: `getValue`, `setValue`, `asObservable` (to
transform it into an Observable), and `finish` (completes all
observables that have been subscribed to).

An `ISharedReference` can simply be created through the
`createSharedReference` function, which only takes one arguement: the
initial value it should be set to.

An `IReadOnlySharedReference` interface was also added. It is exactly the
same thing than a `ISharedReference` but without the `setValue` and
`finish` methods, thus it only has the possibility to read into the
value.

The idea is that an `ISharedReference` is completely compatible and thus
upcastable at any time into a `IReadOnlySharedReference`.

Here, modules with read+write capabilities could require an
`ISharedReference` whereas other modules who can only read it can ask
for a `IReadOnlySharedReference` instead. The same object will be passed
in the two but TypeScript will enforce that only those in possesion of
an `ISharedReference` can mutate it.

Moreover, the `asObservable` method is clearly documented and
implemented as emitting the initial value first, a semantic which is
lost when we upcasted to an Observable.
This behavior is thus both enforced and documented through typing.

---

I'm however conscious that bringing custom concepts can also be a
burden, even when the idea is to remove it (the burden).

That's why I'm eagerly awaiting for your input and if you think whether
this makes sense and is better than what we have today.