ttml: Reinterpret percentages in fontSize attribute
Previously when the TTML fontSize styling attribute was parsed, the
RxPlayer made as if percentages keep the same semantics between
TTML and CSS.

However it isn't at all:
  - [In
    CSS](https://drafts.csswg.org/css-fonts/#valdef-font-size-length-percentage),
    a percentage indicated as a length indicates the font-size relative to
    the one of its parent.
  - [In
    TTML](https://www.w3.org/TR/2018/REC-ttml1-20181108/#style-attribute-fontSize),
    it seems to be in most cases the percentage relaive to the Computed
    Cell Size.

So if I understood correctly all this, it is basically the equivalent of
the `c` unit, multiplied by 100.

This also seems to be how other player are interpreting it:
  - Shaka-player:
https://github.com/google/shaka-player/blob/7c81b0b3f779bcec90b58afce332935d2f2670cf/lib/text/ui_text_displayer.js#L534-L541

  - dash.js:
https://github.com/Dash-Industry-Forum/dash.js/blob/a2b96237861154434dcb8bb572d8563c1437a1f4/src/streaming/text/TextTracks.js#L319-L323

Though there is another part to the TTML specification I do not fully
understand yet:
`if not region element, then relative to the parent element's computed font size`

Does this mean that if the cue has no linked `region` element, it is the
same definition than for CSS? I don't know, but if that's the case, it is
not properly handled now.