api: re-switch to SEEKING state instead of BUFFERING when seeking to already-buffered data
A small regression was brought by the v3.26.1, introduced by the new
ability to detect when playback is "frozen" (so we can then un-freeze it).
Basically, seeking to already-buffered data in the buffer could lead to
an incorrect very brief "BUFFERING" state instead of a "SEEKING" state.

This happens because the video element could pause playback with a
`readyState` set to `1` a brief moment after seeking even when it is to
already-buffered data.

Previously, when playback was paused due to either a low `readyState` or
due to not having enough buffer, the clock (a structure internal to the
RxPlayer which emits playback conditions regularly) emitted a `stalled`
status.
Depending on if the media element is seeking when this happens, that
state was then translated to either the "SEEKING" state or the
"BUFFERING" state by the API.

But the freezing detection in the v3.26.1 changed how this works.
Now, the `stalled` status has been split in two:
  - the `rebuffering` status, for when we are paused due to not having
    enough buffer
  - the `freezing` status, for when we are paused due to unknown reasons

A `freezing` status always led to a "BUFFERING" state and we never
checked in that case if the media element was seeking.

This is why seeking to already-buffered data led to a `"BUFFERING"` state
being emitted: we do have enough buffer but the briefly low `readyState`
indicated that playback was paused. So we had a `freezing` status and no
`rebuffering` status and so we had a `"BUFFERING"` state.

But we could not just check the `seeking` property of the media element
and emit `"SEEKING"` instead accordingly, as we have to also take into
consideration what we call "internal seeks", which are seeks performed
internally by the RxPlayer.

For the latter, we do want to emit the `"BUFFERING"` state, thus
reserving the "SEEKING" state for seeking actions the application is
aware of (i.e. when it is the one asking for it).

This was already handled previously for the `stalled` state and it is
still now for the `rebuffering` state, but it wasn't for the `freezing`
state.
This commit thus implements a logic compatible to both.

Though we may need to start a discussion on what to do with those brief
`readyState` set to `1` after an internal-seek, considering that we may
not event want to send the `"BUFFERING"` state in that case.